Moved the Reset Grid function and button from Grid component to the PatternForm component.
