+++ components/Grid.tsx
@@ -1,6 +1,4 @@
-import Button from "@/components/Button";
-
-// todos
+// TODO
 // allow user to create their own grid with custom size
 // allow user to name the pattern and save to their account
 
@@ -9,23 +7,9 @@ import Button from "@/components/Button";
 // add a paint dropper tool
 // allow selecting color with hex code/manually
 
-// allow click to fill a pixel
-// allow draggin to fill multiple pixels on the grid
+// allow dragging to fill multiple pixels on the grid
 function Grid({ children }: { children: JSX.Element[] }) {
   // TODO: split grid into rows and cols
-
-  function handleResetGrid(e: React.MouseEvent) {
-    e.preventDefault();
-    const target = e.target as HTMLInputElement;
-
-    let pixels = document.querySelectorAll(".grid-pixel");
-    pixels &&
-      pixels.forEach((p) => {
-        p.classList.remove("bg-green-700");
-        p.classList.remove("text-green-700");
-        p.classList.add("text-white");
-      });
-  }
 
   return (
     <>
@@ -33,7 +17,6 @@ function Grid({ children }: { children: JSX.Element[] }) {
         <div className=" border-solid border-2 grid grid-cols-12">
           {children}
         </div>
-        <Button handleClick={handleResetGrid} buttonText="Reset Grid" />
       </div>
     </>
   );
